update dir store
