+++ src/App.tsx
@@ -3,7 +3,7 @@ import './App.css'
 import { Outlet } from 'react-router-dom'
 import Header from './components/Header/Header'
 import { Provider } from 'react-redux'
-import { store } from './app/store'
+import { store } from './app/store/store'
 import SearchBar from './components/SearchBar/SearchBar'
 
 const App = () => {
