+++ src/utils/test-utils.tsx
@@ -3,8 +3,8 @@ import { render } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import type { PropsWithChildren, ReactElement } from 'react'
 import { Provider } from 'react-redux'
-import type { AppStore, RootState } from '../app/store'
-import { makeStore } from '../app/store'
+import type { AppStore, RootState } from '../app/store/store'
+import { makeStore } from '../app/store/store'
 
 /**
  * This type extends the default options for
